Add a stub endpoint for receiving provider event data
* Add a stub endpoint for receiving provider event data
* Add swagger generator to produce api spec from  docs
* Add swagger endpoint for swagger editor ui

Use existing profile table after talking to register-and-partner devs.
The events table will likely gain another FK to mentors at some point but that profile doesn't exist yet.

ParticipantProfile model with states and transitions
ParticipantEvent model for paper_trail event change storage. May want to add in the json blob for this so it can be fully replayed or evidenced.

Co-authored-by: Pavel Lisovin <pavel.lisovin@digital.education.gov.uk>
Co-authored-by: Paul McKibbin <pmckibbin@gmail.com>